+++ NEWS.md
@@ -2,5 +2,11 @@
 
 * Initial release
 
+# ShinyTester 0.2.0 (in development)
+
+* Added parameters to ShinyHierarchy: showInternalFunctions, showCommentedOutChunks parameters, and fixed a bug in Chunking
+* ShinyHierarchy now is able to ouptut a Sankey Chart as well
+* Added new function: ProcessReactives... not operational yet.
+